Refactoring: merge request/response functionality into call
